Refactor: create feature-based component folders
