+++ src/App.jsx
@@ -1,8 +1,8 @@
 import { useState } from 'react'
 import './App.css'
-import Header from './Header.jsx'
-import Barner from './Barner.jsx'
-import Nav from './Nav.jsx'
+import Header from './components/layout/Header.jsx'
+import Barner from './components/layout/Barner.jsx'
+import Nav from './components/layout/Nav.jsx'
 
 function App() {
