--- src/Navbar.jsx
@@ -1,22 +1,0 @@
-import React from 'react'
-
-function Navbar() {
-  return (
-    <div>
-        <nav className="flex justify-between ">
-            <div>
-              <span>malakus <span>&gt;</span></span>
-            </div>
-
-            <div className="flex gap-4">
-              <a href="">Thoughts</a>
-              <a href="">Experiments</a>
-              <a href="">Work</a>
-            </div>
-        </nav>
-
-    </div>
-  )
-}
-
-export default Navbar
